Adapt data types for MS-SQL

+++ src/main/resources/db/migration/V1_1__CreateSchema.sql
@@ -2,7 +2,7 @@ create table INVENTORY (
     itemId varchar(255) not null,
     link varchar(255),
     location varchar(255),
-    quantity int4 not null,
+    quantity int not null,
     primary key (itemId)
 );
 
@@ -10,6 +10,6 @@ create table PRODUCT_CATALOG (
     itemId varchar(255) not null,
     description text,
     name varchar(255),
-    price float8 not null,
+    price float not null,
     primary key (itemId)
 );
